requirements.txt
carbontracker==1.1.6
einops==0.4.0
librosa==0.8.1
numpy==1.20.3
pytorch_lightning==1.5.9
PyYAML==6.0
torch==1.11.0
tensorboard==2.8.0
torchaudio==0.11.0
tqdm==4.62.3
git+https://github.com/caillonantoine/cached_conv.git@v2.3.5#egg=cached_conv
git+https://github.com/caillonantoine/UDLS.git@v1.0.0#egg=udls
git+https://github.com/aliutkus/speechmetrics#egg=speechmetrics[gpu]
git+https://github.com/microsoft/DeepSpeed.git@6719b46bd8792755cfc75eaa448bf7fa65fd99eb